requirements.txt
tabulate
nfl_data_py
flake8
google-generativeai
python-dotenv
pandas
espn_api
pyyaml
markdown
requests
openai
lxml